Prioritize fill-in-middle mode if appropriate
Ollama's API provides a `suffix` parameter that is passed through to
prompt templates, without which some models[1] do not appropriately
return fill-in-middle output even with the model-specific tokens in the
prompt. This ensures that if we are in a FIM context (i.e. where
`starhugger--fim-concatenate` is used) we pass that parameter.

1: Definitely true for the default templates from Ollama for QwenCoder2.5